# app/static/swagger.yaml
swagger: "2.0"
info:
  title: "Mechanic Shop API"
  description: "A comprehensive RESTful API built with Flask for managing mechanic shop operations, featuring token authentication, rate limiting, caching, customer management, mechanic records, service ticket tracking, and inventory management."
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
paths:
  /customers/login:
    post:
      tags:
        - "Customers"
      summary: "Login (generate token)"
      description: "Validates customer credentials and returns a JWT token."
      parameters:
        - in: "body"
          name: "body"
          description: "Login credentials"
          required: true
          schema:
            $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: "Successfully logged in"
          schema:
            $ref: "#/definitions/LoginResponse"
        400:
          description: "Validation error"
        401:
          description: "Invalid email or password"
  /customers:
    post:
      tags:
        - "Customers"
      summary: "Create a customer"
      description: "Create/register a new customer. Email must be unique. Rate limited to 5 per day."
      parameters:
        - in: "body"
          name: "body"
          description: "Customer data"
          required: true
          schema:
            $ref: "#/definitions/CustomerCreatePayload"
      responses:
        201:
          description: "Customer created"
          schema:
            $ref: "#/definitions/Customer"
        400:
          description: "Validation error or duplicate email"
    get:
      tags:
        - "Customers"
      summary: "Get customers (paginated)"
      description: "Get all customers with pagination (cached 60 seconds)."
      parameters:
        - in: "query"
          name: "page"
          type: "integer"
          required: false
          description: "Page number (default 1)"
        - in: "query"
          name: "per_page"
          type: "integer"
          required: false
          description: "Items per page (default 10)"
      responses:
        200:
          description: "Customers retrieved"
          schema:
            $ref: "#/definitions/PaginatedCustomers"
  /customers/{customer_id}:
    get:
      tags:
        - "Customers"
      summary: "Get a customer"
      description: "Get a specific customer by ID."
      parameters:
        - in: "path"
          name: "customer_id"
          required: true
          type: "integer"
          description: "Customer ID"
      responses:
        200:
          description: "Customer found"
          schema:
            $ref: "#/definitions/Customer"
        404:
          description: "Customer not found"
    put:
      tags:
        - "Customers"
      summary: "Update a customer"
      description: "Update a specific customer. Protected by token and rate limited (5 per month)."
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "customer_id"
          required: true
          type: "integer"
          description: "Customer ID"
        - in: "body"
          name: "body"
          description: "Customer data to update"
          required: true
          schema:
            $ref: "#/definitions/CustomerUpdatePayload"
      responses:
        200:
          description: "Customer updated"
          schema:
            $ref: "#/definitions/Customer"
        400:
          description: "Validation error"
        404:
          description: "Customer not found"
        401:
          description: "Missing/invalid token"
    delete:
      tags:
        - "Customers"
      summary: "Delete a customer"
      description: "Delete a specific customer by ID. Protected by token and rate limited (5 per day)."
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "customer_id"
          required: true
          type: "integer"
          description: "Customer ID"
      responses:
        200:
          description: "Customer deleted"
          schema:
            $ref: "#/definitions/DeleteCustomerResponse"
        404:
          description: "Customer not found"
        401:
          description: "Missing/invalid token"
...
definitions:
  LoginCredentials:
    type: "object"
    properties:
      email:
        type: "string"
      password:
        type: "string"
    required:
      - email
      - password
  LoginResponse:
    type: "object"
    properties:
      token:
        type: "string"
      message:
        type: "string"
      status:
        type: "string"
  Customer:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
  CustomerCreatePayload:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"
    required:
      - name
      - email
      - phone
      - password
  CustomerUpdatePayload:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"
  DeleteCustomerResponse:
    type: "object"
    properties:
      message:
        type: "string"
  PaginatedCustomers:
    type: "object"
    properties:
      customers:
        type: "array"
        items:
          $ref: "#/definitions/Customer"
      page:
        type: "integer"
      per_page:
        type: "integer"
      total_pages:
        type: "integer"
      total_items:
        type: "integer"
      has_next:
        type: "boolean"
      has_prev:
        type: "boolean"
      next_page:
        type: "integer"
        x-nullable: true
      prev_page:
        type: "integer"
        x-nullable: true
...